AdminServer: Move drone id into URL for enter request path

+++ src/main/java/it/cutecchia/sdp/admin/server/stores/DronesStore.java
@@ -1,8 +1,6 @@
 package it.cutecchia.sdp.admin.server.stores;
 
 import it.cutecchia.sdp.admin.server.beans.DroneInfo;
-
-import java.net.InetSocketAddress;
 import java.util.Set;
 
 public interface DronesStore {
@@ -10,9 +8,10 @@ public interface DronesStore {
 
   class DroneIdNotFound extends Exception {}
 
-  DroneInfo addNewDrone(long droneId, String ipAddress, int connectionPort) throws DroneIdAlreadyInUse;
+  DroneInfo addNewDrone(int droneId, String ipAddress, int connectionPort)
+      throws DroneIdAlreadyInUse;
 
-  void removeDroneById(long droneId) throws DroneIdNotFound;
+  void removeDroneById(int droneId) throws DroneIdNotFound;
 
   Set<DroneInfo> getRegisteredDrones();
 }
